# api/openapi.yaml
openapi: 3.0.3
info:
  title: OurSay Account API
  description: Account registration & authentication. Email OTP is
    bootstrap/recovery only; WebAuthn passkeys are the day-to-day login. These
    passkeys are the ACCOUNT-LOGIN factor and are separate from the civic
    thread-signing keys in @oursay/public-record.
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token
    cookieAuth:
      type: apiKey
      in: cookie
      name: oursay_session
...
paths:
...
  /v1/me/jurisdictions:
    get:
      summary: List the authenticated user's jurisdiction memberships
      tags:
        - me
      security: *a1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  jurisdictionIds:
                    type: array
                    items:
                      type: string
                required:
                  - jurisdictionIds
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
    put:
      summary: Replace jurisdiction subscriptions (oursay-global is always retained;
        roles are never changed)
      tags:
        - me
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                jurisdictionIds:
                  type: array
                  items:
                    type: string
              required:
                - jurisdictionIds
      security: *a1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  jurisdictionIds:
                    type: array
                    items:
                      type: string
                required:
                  - jurisdictionIds
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
  /v1/me/signing-prefs:
    get:
      summary: Get per-action signing preferences
      tags:
        - me
      security: *a1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                  enum:
                    - quick
                    - ask
                    - passkey
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
    patch:
      summary: Merge signing preferences (null deletes a key)
      tags:
        - me
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                oneOf:
                  - type: string
                    enum:
                      - quick
                      - ask
                      - passkey
                  - type: "null"
      security: *a1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                  enum:
                    - quick
                    - ask
                    - passkey
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
  /v1/me/visibility:
    get:
      summary: Get the account-default author visibility
      tags:
        - me
      security: *a1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  visibility:
                    type: string
                    enum:
                      - anonymous
                      - my_officials
                      - all_officials
                      - my_district
                      - my_jurisdiction
                      - id_verified
                      - public
                required:
                  - visibility
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
    patch:
      summary: Set the account-default author visibility
      tags:
        - me
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                visibility:
                  type: string
                  enum:
                    - anonymous
                    - my_officials
                    - all_officials
                    - my_district
                    - my_jurisdiction
                    - id_verified
                    - public
              required:
                - visibility
      security: *a1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  visibility:
                    type: string
                    enum:
                      - anonymous
                      - my_officials
                      - all_officials
                      - my_district
                      - my_jurisdiction
                      - id_verified
                      - public
                required:
                  - visibility
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
  /v1/me/threads/{threadId}/visibility:
    put:
      summary: Set or clear a per-thread visibility override (null clears)
      tags:
        - me
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                visibility:
                  oneOf:
                    - type: string
                      enum:
                        - anonymous
                        - my_officials
                        - all_officials
                        - my_district
                        - my_jurisdiction
                        - id_verified
                        - public
                    - type: "null"
              required:
                - visibility
      parameters:
        - schema:
            type: string
          in: path
          name: threadId
          required: true
      security: *a1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  visibility:
                    oneOf:
                      - type: string
                        enum:
                          - anonymous
                          - my_officials
                          - all_officials
                          - my_district
                          - my_jurisdiction
                          - id_verified
                          - public
                      - type: "null"
                required:
                  - visibility
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
  /v1/me/districts:
    get:
      summary: Home district slugs for the authenticated viewer (self-only raw slugs)
      tags:
        - me
      security: *a1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  districts:
                    type: array
                    items:
                      type: string
                required:
                  - districts
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
...
  /v1/me/shares/{shareKey}:
    post:
      summary: Record a share mark for deduplicated share counts
      tags:
        - me
      parameters:
        - schema:
            type: string
          in: path
          name: shareKey
          required: true
      security: *a1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  counted:
                    type: boolean
                  count:
                    type: number
                required:
                  - counted
                  - count
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
...